Repositories needed for a simple Library application
I wanted to learn Spring Data JPA. So I have come up with a scenario and
tried to implement some JPA repositories.

At this stage only the repositories are ready. I will keep improving
this project with more layers as and when I learn.

Some things to keep in mind :
- this application definitely doesn't target a huge volume library. this
is an inspiration that occurred when the management where I work had to
maintain a hand written register to keep a log of the books lended from
the library which hardly circulates 100 books.

Features :
* There can be multiple books for the same isbn
* Books can be contributed by employees/management. they can be
contributed as part of lost book penalty as well
* Penalty for damaged book, lost book and missing due date will be
designed
* Rating and reviews for book can be provided by users - not necessary
that they should have read it from this library. If they have already
read it somewhere they can enrich the reviews here
* Watchers - users can watch for a book to become available
* Subscribers - users can subscribe for new book additions

Again, the focus of this task is not to build a rich application. It is
to try and attempt all the basic JPA functions
- primary keys
- composite keys with reference keys
- enumerations
- query dsl
- transient columns
- join columns