More changes done to JPA repositories
I am currently focussing on the wrapper layer for the Book Catalog and
Book
There are so many changes at JPA - maybe because I do not know the depth
of Query DSL. I am taking help from the wrapper to do some of the
feature implementations.
Will have to revisit some cases.
like how can I avoid entityManager.clear()
I thought it was just a test thing. but maybe there is something to
automatically update the catalog with the newly added books and I just
do not know it.

Also, I could have easily written an sql script to update the test data.
maybe if i did that, the IT would not look so clumsy. But then without
setting up data using the services themselves, I wouldn't have caught
the entity manager to be cleared for the join columns to be updated.
I will write one test case using SQL grouping maybe later